# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Theorema Aureum 143 — Certificate Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CertificateStatus = typeof CertificateStatus[keyof typeof CertificateStatus];


export const CertificateStatus = {
  CERTIFIED: 'CERTIFIED',
  AWAITING: 'AWAITING',
  LOCKED: 'LOCKED',
  DISCHARGED: 'DISCHARGED',
} as const;

export interface Certificate {
  /** e.g. M1, M2, ... M7 */
  moduleId: string;
  title: string;
  /** One-sentence summary of what this module certifies */
  claim: string;
  status: CertificateStatus;
  /** @nullable */
  sourceFile?: string | null;
  sourceSha: string;
  stdoutSha: string;
  /** SHA-256 digests of parent module outputs this module depends on */
  parentShas: string[];
  /** Position in the M1→M7 DAG (1-indexed) */
  dagPosition: number;
  /**
     * Object storage path of the uploaded PDF
     * @nullable
     */
  pdfObjectPath?: string | null;
  /**
     * Lean theorem name bound to this certificate
     * @nullable
     */
  leanBinding?: string | null;
  /** @nullable */
  notes?: string | null;
  updatedAt?: string;
}

export interface CertificateUpdate {
  /** @nullable */
  pdfObjectPath?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  status?: string | null;
}

export interface CertificateSummary {
  /** Master manifest SHA-256 (cat m1.out..m6.out | sha256sum) */
  masterSha: string;
  totalModules: number;
  certifiedCount: number;
  awaitingCount: number;
  dagSealed: boolean;
  pdfUploadedCount?: number;
  pdfTotal?: number;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}