# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Theorema Aureum 143 — Certificate Ledger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /certificates/{moduleId}:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get a specific certificate by module ID (e.g. M1)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "404":
          description: Not found
    patch:
      operationId: updateCertificate
      tags: [certificates]
      summary: Update a certificate (attach PDF, notes, or mark verified)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CertificateUpdate"
      responses:
        "200":
          description: Updated certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "404":
          description: Not found
...
components:
  schemas:
...
    Certificate:
      type: object
      required: [moduleId, title, claim, status, sourceSha, stdoutSha, parentShas, dagPosition]
      properties:
        moduleId:
          type: string
          description: "e.g. M1, M2, ... M7"
        title:
          type: string
        claim:
          type: string
          description: One-sentence summary of what this module certifies
        status:
          type: string
          enum: [CERTIFIED, AWAITING, LOCKED, DISCHARGED]
        sourceFile:
          type: ["string", "null"]
        sourceSha:
          type: string
        stdoutSha:
          type: string
        parentShas:
          type: array
          items:
            type: string
          description: SHA-256 digests of parent module outputs this module depends on
        dagPosition:
          type: integer
          description: Position in the M1→M7 DAG (1-indexed)
        pdfObjectPath:
          type: ["string", "null"]
          description: Object storage path of the uploaded PDF
        leanBinding:
          type: ["string", "null"]
          description: Lean theorem name bound to this certificate
        notes:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
    CertificateUpdate:
      type: object
      properties:
        pdfObjectPath:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
...